Minor improvements, mostly output-related

+++ src/main/java/dk/magenta/databroker/cprvejregister/dataproviders/registers/LokalitetsRegister.java
@@ -118,7 +118,7 @@ public class LokalitetsRegister extends CprSubRegister {
 
     protected void saveRunToDatabase(RegisterRun run, RegistreringInfo registreringInfo) {
         this.log.info("Storing LokalitetEntities in database");
-        long time = this.indepTic();
+        double time = this.indepTic();
         ModelUpdateCounter counter = new ModelUpdateCounter();
         counter.setLog(this.log);
         LokalitetRegisterRun lrun = (LokalitetRegisterRun) run;
@@ -140,7 +140,7 @@ public class LokalitetsRegister extends CprSubRegister {
         counter.printFinalEntriesProcessed();
         int count = counter.getCount();
         time = this.toc(time);
-        this.log.info(count + " LokalitetEntities stored in " + time + " ms (avg " + ((double)time / (double)count) + " ms)");
+        this.log.info(count + " LokalitetEntities stored in " + time + " ms (avg " + (time / (double)count) + " ms)");
         registreringInfo.logProcess(this.log);
     }
