+++ src/main/java/dk/magenta/databroker/cprvejregister/dataproviders/registers/PostnummerRegister.java
@@ -144,7 +144,7 @@ public class PostnummerRegister extends CprSubRegister {
 
     protected void saveRunToDatabase(RegisterRun run, RegistreringInfo registreringInfo) {
         this.log.info("Storing PostnummerEntities in database");
-        long time = this.indepTic();
+        double time = this.indepTic();
         ModelUpdateCounter counter = new ModelUpdateCounter();
         counter.setLog(this.log);
         PostnummerRegisterRun prun = (PostnummerRegisterRun) run;
@@ -173,7 +173,7 @@ public class PostnummerRegister extends CprSubRegister {
         counter.printFinalEntriesProcessed();
         int count = counter.getCount();
         time = this.toc(time);
-        this.log.info(count + " PostnummerEntities stored in " + time + " ms (avg " + ((double)time / (double)count) + " ms)");
+        this.log.info(count + " PostnummerEntities stored in " + time + " ms (avg " + (time / (double)count) + " ms)");
         registreringInfo.logProcess(this.log);
     }
